Add: components/domain 폴더 추가 도메인과 밀접한 components 위치
