+++ src/components/domains/detail/CommentCountView/CommentCountView.module.scss
@@ -1,0 +1,18 @@
+.footer_item_list_wrap {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  margin: 0 5%;
+  gap: 7px;
+  padding-bottom: 35px;
+  .footer_item_wrap {
+    display: flex;
+    align-items: center;
+    gap: 5px;
+    & > img {
+      width: 22px;
+      height: 22px;
+      cursor: default;
+    }
+  }
+}
