+++ src/components/domains/detail/CommentSection/commentSection.module.scss
@@ -1,0 +1,43 @@
+.comment_wrap {
+  .comment_list_wrap {
+    margin-bottom: 30px;
+  }
+
+  .comment_textarea {
+    width: 100%;
+    height: 150px;
+    padding: 10px;
+    border: 2px solid #c8c5cb;
+    border-radius: 10px;
+    outline: none;
+
+    &:focus {
+      border: 2px solid #98b487;
+    }
+  }
+
+  .comment_button_wrap {
+    display: flex;
+    gap: 15px;
+    width: 250px;
+    margin-left: auto;
+    margin-top: 25px;
+
+    .cancle_button {
+      border: 0;
+      border-radius: 8px;
+      padding: 13px 20px;
+      font-size: 16px;
+      font-weight: 700;
+      background-color: #c0c0c0;
+      color: #fff;
+    }
+  }
+  .comment_input_wrap {
+    .comment_input {
+      border: 1px solid green;
+      width: 100%;
+      padding: 18px 44px;
+    }
+  }
+}
