+++ src/components/domains/detail/SectionContent/sectionContent.module.scss
@@ -1,0 +1,10 @@
+.content_wrap {
+  min-height: 680px;
+  padding: 20px 5%;
+}
+
+.detail_line {
+  border: 1px solid #eef3eb;
+  width: 100%;
+  margin: 20px auto;
+}
