+++ src/components/domains/detail/SectionTitle/sectionTitle.module.scss
@@ -1,0 +1,40 @@
+.header_title_wrap {
+  display: flex;
+  justify-content: space-between;
+  margin-left: 5%;
+  margin-right: 5%;
+  font-family: Noto Sans KR;
+  font-size: 35px;
+  font-weight: 500;
+  .title_left {
+    max-width: 50vw;
+    & h1 {
+      width: 100%;
+      height: 100%;
+      overflow: hidden;
+      text-overflow: ellipsis;
+      white-space: nowrap;
+      // word-wrap: break-word;
+    }
+  }
+  .title_right {
+    display: flex;
+    gap: 10px;
+    .like_icon {
+      width: 35px;
+      height: 35px;
+      transition: 0.3s;
+      &:hover {
+        opacity: 0.8;
+        transform: scale(1.1);
+        transition: 0.3s;
+      }
+      &:active {
+        // width: 100px;
+        opacity: 1;
+        transform: scale(1.2);
+        transition: 10ms;
+      }
+    }
+  }
+}
