+++ src/components/domains/detail/index.ts
@@ -1,0 +1,9 @@
+// import { SectionCategory } from './SectionCategory/SectionCategory';
+
+export * from './SectionTitle/SectionTitle';
+export * from './SectionCategory/SectionCategory';
+export * from './SectionContactView/SectionContactView';
+export * from './SectionContent/SectionContent';
+export * from './CommentButtons/CommentButtons';
+export * from './CommentCountView/CommentCountView';
+export * from './CommentSection/index';
