+++ src/pages/DetailPage/DetailPage.module.scss
@@ -56,81 +56,81 @@
 }
 
 // detail
-.header_title_wrap {
-  display: flex;
-  justify-content: space-between;
-  margin-left: 5%;
-  margin-right: 5%;
-  font-family: Noto Sans KR;
-  font-size: 35px;
-  font-weight: 500;
-  .title_left {
-    max-width: 50vw;
-    & h1 {
-      width: 100%;
-      height: 100%;
-      overflow: hidden;
-      text-overflow: ellipsis;
-      white-space: nowrap;
-      // word-wrap: break-word;
-    }
-  }
-  .title_right {
-    display: flex;
-    gap: 10px;
-    .like_icon {
-      width: 35px;
-      height: 35px;
-      transition: 0.3s;
-      &:hover {
-        opacity: 0.8;
-        transform: scale(1.1);
-        transition: 0.3s;
-      }
-      &:active {
-        // width: 100px;
-        opacity: 1;
-        transform: scale(1.2);
-        transition: 10ms;
-      }
-    }
-  }
-}
+// .header_title_wrap {
+//   display: flex;
+//   justify-content: space-between;
+//   margin-left: 5%;
+//   margin-right: 5%;
+//   font-family: Noto Sans KR;
+//   font-size: 35px;
+//   font-weight: 500;
+//   .title_left {
+//     max-width: 50vw;
+//     & h1 {
+//       width: 100%;
+//       height: 100%;
+//       overflow: hidden;
+//       text-overflow: ellipsis;
+//       white-space: nowrap;
+//       // word-wrap: break-word;
+//     }
+//   }
+//   .title_right {
+//     display: flex;
+//     gap: 10px;
+//     .like_icon {
+//       width: 35px;
+//       height: 35px;
+//       transition: 0.3s;
+//       &:hover {
+//         opacity: 0.8;
+//         transform: scale(1.1);
+//         transition: 0.3s;
+//       }
+//       &:active {
+//         // width: 100px;
+//         opacity: 1;
+//         transform: scale(1.2);
+//         transition: 10ms;
+//       }
+//     }
+//   }
+// }
 
-.contact_view_wrap {
-  margin: 30px 5%;
-  & p {
-    & a {
-      color: blue;
-      text-decoration-line: underline;
-    }
-  }
-}
+// .contact_view_wrap {
+//   margin: 30px 5%;
+//   & p {
+//     & a {
+//       color: blue;
+//       text-decoration-line: underline;
+//     }
+//   }
+// }
 
-.category_wrap {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  font-weight: 700;
-  margin-left: 5%;
-  margin-right: 5%;
-  .category_left {
-    display: flex;
-    gap: 10px;
-    .category_item {
-      width: 100px;
-      height: 40px;
-      border-radius: 8px;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      border: 1.5px solid #aac49a;
+// .category_wrap {
+//   display: flex;
+//   align-items: center;
+//   justify-content: space-between;
+//   font-weight: 700;
+//   margin-left: 5%;
+//   margin-right: 5%;
+//   .category_left {
+//     display: flex;
+//     gap: 10px;
+//     .category_item {
+//       width: 100px;
+//       height: 40px;
+//       border-radius: 8px;
+//       display: flex;
+//       justify-content: center;
+//       align-items: center;
+//       border: 1.5px solid #aac49a;
 
-      color: #568a35;
-    }
-    color: #568a35;
-  }
-}
+//       color: #568a35;
+//     }
+//     color: #568a35;
+//   }
+// }
 
 .category_right {
   .status_button {
@@ -164,10 +164,10 @@
   }
 }
 
-.content_wrap {
-  min-height: 680px;
-  padding: 20px 5%;
-}
+// .content_wrap {
+//   min-height: 680px;
+//   padding: 20px 5%;
+// }
 
 .detail_line {
   border: 1px solid #eef3eb;
@@ -175,65 +175,21 @@
   margin: 20px auto;
 }
 
-.footer_item_list_wrap {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  margin: 0 5%;
-  gap: 7px;
-  padding-bottom: 35px;
-  .footer_item_wrap {
-    display: flex;
-    align-items: center;
-    gap: 5px;
-    & > img {
-      width: 22px;
-      height: 22px;
-      cursor: default;
-    }
-  }
-}
-
-.comment_wrap {
-  .comment_list_wrap {
-    margin-bottom: 30px;
-  }
-
-  .comment_textarea {
-    width: 100%;
-    height: 150px;
-    padding: 10px;
-    border: 2px solid #c8c5cb;
-    border-radius: 10px;
-    outline: none;
-
-    &:focus {
-      border: 2px solid #98b487;
-    }
-  }
-
-  .comment_button_wrap {
-    display: flex;
-    gap: 15px;
-    width: 250px;
-    margin-left: auto;
-    margin-top: 25px;
-
-    .cancle_button {
-      border: 0;
-      border-radius: 8px;
-      padding: 13px 20px;
-      font-size: 16px;
-      font-weight: 700;
-      background-color: #c0c0c0;
-      color: #fff;
-    }
-  }
-  .comment_input_wrap {
-    .comment_input {
-      border: 1px solid green;
-      width: 100%;
-      padding: 18px 44px;
-    }
-  }
-}
+// .footer_item_list_wrap {
+//   display: flex;
+//   justify-content: flex-end;
+//   align-items: center;
+//   margin: 0 5%;
+//   gap: 7px;
+//   padding-bottom: 35px;
+//   .footer_item_wrap {
+//     display: flex;
+//     align-items: center;
+//     gap: 5px;
+//     & > img {
+//       width: 22px;
+//       height: 22px;
+//       cursor: default;
+//     }
+//   }
+// }
